refactor: move logging functionality ot a separate folder
